add referencing image for the TerraLand
TerraLand is implemented as a 9 x 13 Terrain Array.  In the implementation, each column has a type which can be one of the 7 terrain types or river. e.g the second row in the map consists of terrains that have the following types in the implementation : DESERT, RIVER, RIVER, PLAINS, SWAMP, RIVER, RIVER, DESERT, SWAMP, RIVER, RIVER, DESERT, RIVER.